MarketPriceActivity to MarketPriceFragment
